扩展grpc订阅功能

+++ system/dapp/commands/block.go
@@ -363,7 +363,7 @@ func addPushSubscribe(cmd *cobra.Command, args []string) {
 		LastSequence:  lastSeq,
 		LastHeight:    lastHeight,
 		LastBlockHash: lastBlockHash,
-		Type:          pushType,
+		Type: int32(pushType),
 	}
 
 	var res types.ReplySubscribePush
